+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Models/ResourceMetadata.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 namespace Azure.Generator.Management.Models
 {
     internal record ResourceMetadata(
+        string ResourceIdPattern,
         string ResourceType,
         ResourceScope ResourceScope,
         IReadOnlyList<ResourceMethod> Methods,