+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Providers/TagMethodProviders/BaseTagMethodProvider.cs
@@ -91,7 +91,7 @@ namespace Azure.Generator.Management.Providers.TagMethodProviders
             foreach (var (kind, method) in resourceClientProvider.ResourceServiceMethods)
             {
                 var operation = method.Operation;
-                if (kind == Models.OperationKind.Get)
+                if (kind == Models.ResourceOperationKind.Get)
                 {
                     getServiceMethod = method;
                     break;